pyproject.toml
[project]
name = "hydroportal"
version = "0.1.0"
description = "HydroPortal 水网门户 - Unified Web portal for HydroMind ecosystem"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "pydantic>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=4.0,<4.2",
    "httpx>=0.24",
    "websockets>=11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24",
    "pytest-asyncio>=0.21",
    "starlette[full]>=0.27",
]
full = [
    "hydromind-contracts",
    "hydroguard",
    "hydrodesign",
    "hydrolab",
    "hydroedu",
    "hydroarena",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
